Put Swagger attributes into own folder and inlined the Startup partial class
