+++ src/Speakr.TalksApi/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using Microsoft.AspNetCore.Hosting;
+using Speakr.TalksApi.AppStart;
 
 namespace Speakr.TalksApi
 {
